Tests/LibWeb/Text/expected/css/adopted-stylesheet-pending-invalidation-survives-adoption.txt

initial color: rgb(255, 0, 0)
after adoption color: rgb(0, 128, 0)
